pyproject.toml
[tool.poetry]
name = "dijksta-tests"
version = "0.1.0"
description = ""
authors = ["Bruno Toffoli <btoffoli@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
sklearn2 = "^0.0.13"
scipy = "^1.6.1"
scipy-extra = "^0.1"
networkx = "^2.6.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
